:recycle: Migrate fluent.js compatibility tests to RSpec
Convert standalone rake-based compatibility testing to proper RSpec
integration where each fixture becomes an individual test example.

- Add spec/support/fluent_js/ with AstComparator, FixtureLoader, TestHelper
- Create spec/fluent_js_compatibility_spec.rb with 98 individual examples
- Remove compat/ directory and lib/tasks/compatibility.rake
- Remove .github/workflows/compatibility.yml
- Update .rubocop.yml, README.md, CLAUDE.md, AGENTS.md
