--- .github/workflows/compatibility.yml
@@ -1,90 +1,0 @@
-name: Compatibility Tests
-
-# This workflow runs compatibility tests when relevant files change.
-# This keeps the main CI fast while ensuring compatibility is verified when needed.
-
-on:
-  push:
-    branches:
-      - main
-    paths:
-      - 'lib/foxtail/parser/**'
-      - 'lib/foxtail/parser.rb'
-      - 'lib/foxtail/bundle/ast_converter.rb'
-      - 'spec/parser/**'
-      - 'compat/fluent_js/**'
-      - '.github/workflows/compatibility.yml'
-
-  pull_request:
-    paths:
-      - 'lib/foxtail/parser/**'
-      - 'lib/foxtail/parser.rb'
-      - 'lib/foxtail/bundle/ast_converter.rb'
-      - 'spec/parser/**'
-      - 'compat/fluent_js/**'
-      - '.github/workflows/compatibility.yml'
-
-  # Allow manual triggering for comprehensive testing
-  workflow_dispatch:
-
-permissions:
-  contents: read
-  pull-requests: write
-
-jobs:
-  fluent-js:
-    runs-on: ubuntu-latest
-    name: Fluent.js Compatibility
-    if: github.event_name == 'workflow_dispatch' || github.event_name == 'pull_request' || github.event_name == 'push'
-
-    steps:
-      - uses: actions/checkout@v5
-        with:
-          submodules: recursive
-
-      - name: Set up Ruby
-        uses: ruby/setup-ruby@v1
-        with:
-          ruby-version: '3.4'  # Use latest supported version
-          bundler-cache: true
-
-      - name: Generate Fluent.js compatibility report
-        run: |
-          bundle exec rake compatibility:fluentjs
-
-      - name: Upload Fluent.js compatibility report
-        uses: actions/upload-artifact@v4
-        with:
-          name: fluentjs-compatibility-report
-          path: compat/fluentjs_compatibility_report.md
-          retention-days: 30
-
-      - name: Extract Fluent.js compatibility summary
-        if: github.event_name == 'pull_request'
-        id: extract-summary
-        run: |
-          # Read the markdown report
-          REPORT=$(cat compat/fluentjs_compatibility_report.md)
-
-          # Extract the summary table and total fixtures line
-          SUMMARY=$(echo "$REPORT" | sed -n '/## Summary/,/*Total fixtures:/p')
-
-          # Create the comment body
-          {
-            echo 'body<<EOF'
-            echo '## 📊 Fluent.js Compatibility Report'
-            echo ''
-            echo "$SUMMARY"
-            echo ''
-            echo "[Full report available in workflow artifacts](https://github.com/${{ github.repository }}/actions/runs/${{ github.run_id }})"
-            echo ''
-            echo '> 🔍 This compatibility check ran because parser-related files were modified.'
-            echo 'EOF'
-          } >> $GITHUB_OUTPUT
-
-      - name: Comment PR with Fluent.js compatibility summary
-        if: github.event_name == 'pull_request'
-        uses: int128/comment-action@v1
-        with:
-          post: |
-            ${{ steps.extract-summary.outputs.body }}
