+++ spec/support/fluent_js.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+require_relative "fluent_js/ast_comparator"
+require_relative "fluent_js/fixture_loader"
+require_relative "fluent_js/test_helper"
